Minor tidy ups
- Seperate types to their own dedicated files
- Move errors to the top of the file
- Add public initializers (as the types are public, they require their initalizers to be public too)
