+++ Sources/TuistCore/MetadataProviders/SystemFrameworksMetadataProvider.swift
@@ -2,39 +2,27 @@ import Foundation
 import TSCBasic
 import TuistSupport
 
-// MARK: - Metadata
+// MARK: - Provider Errors
 
-public struct SystemFrameworkMetadata: Equatable {
-    var name: String
-    var path: AbsolutePath
-    var status: SDKStatus
-    var source: SDKSource
-}
+public enum SystemFrameworkMetadataProviderError: FatalError, Equatable {
+    case unsupportedSDK(name: String)
+    case unsupportedSDKForPlatform(name: String, platform: Platform)
 
-public enum SDKSource: Equatable {
-    case developer // Platforms/iPhoneOS.platform/Developer/Library
-    case system // Platforms/iPhoneOS.platform/Developer/SDKs/iPhoneOS.sdk/System/Library
-
-    /// Returns the framework search path that should be used in Xcode to locate the SDK.
-    public var frameworkSearchPath: String? {
+    public var description: String {
         switch self {
-        case .developer:
-            return "$(PLATFORM_DIR)/Developer/Library/Frameworks"
-        case .system:
-            return nil
+        case let .unsupportedSDK(sdk):
+            let supportedTypes = SDKType.supportedTypesDescription
+            return "The SDK type of \(sdk) is not currently supported - only \(supportedTypes) are supported."
+        case let .unsupportedSDKForPlatform(name: sdk, platform: platform):
+            return "The SDK \(sdk) is not currently supported on \(platform)."
         }
     }
-}
 
-public enum SDKType: String, CaseIterable, Equatable {
-    case framework
-    case library = "tbd"
-
-    static var supportedTypesDescription: String {
-        let supportedTypes = allCases
-            .map { ".\($0.rawValue)" }
-            .joined(separator: ", ")
-        return "[\(supportedTypes)]"
+    public var type: ErrorType {
+        switch self {
+        case .unsupportedSDK, .unsupportedSDKForPlatform:
+            return .abort
+        }
     }
 }
 
@@ -53,6 +41,8 @@ extension SystemFrameworkMetadataProviding {
 // MARK: - Default Implementation
 
 public final class SystemFrameworkMetadataProvider: SystemFrameworkMetadataProviding {
+    public init() {}
+
     public func loadMetadata(sdkName: String, status: SDKStatus, platform: Platform, source: SDKSource) throws -> SystemFrameworkMetadata {
         let sdkNamePath = AbsolutePath("/\(sdkName)")
         guard let sdkExtension = sdkNamePath.extension,
@@ -92,30 +82,6 @@ public final class SystemFrameworkMetadataProvider: SystemFrameworkMetadataProvi
                     .appending(RelativePath("usr/lib"))
                     .appending(component: name)
             }
-        }
-    }
-}
-
-// MARK: - Errors
-
-public enum SystemFrameworkMetadataProviderError: FatalError, Equatable {
-    case unsupportedSDK(name: String)
-    case unsupportedSDKForPlatform(name: String, platform: Platform)
-
-    public var description: String {
-        switch self {
-        case let .unsupportedSDK(sdk):
-            let supportedTypes = SDKType.supportedTypesDescription
-            return "The SDK type of \(sdk) is not currently supported - only \(supportedTypes) are supported."
-        case let .unsupportedSDKForPlatform(name: sdk, platform: platform):
-            return "The SDK \(sdk) is not currently supported on \(platform)."
-        }
-    }
-
-    public var type: ErrorType {
-        switch self {
-        case .unsupportedSDK, .unsupportedSDKForPlatform:
-            return .abort
         }
     }
 }
