+++ Sources/TuistCore/Models/SDKType.swift
@@ -1,0 +1,13 @@
+import Foundation
+
+public enum SDKType: String, CaseIterable, Equatable {
+    case framework
+    case library = "tbd"
+
+    static var supportedTypesDescription: String {
+        let supportedTypes = allCases
+            .map { ".\($0.rawValue)" }
+            .joined(separator: ", ")
+        return "[\(supportedTypes)]"
+    }
+}
