+++ Sources/TuistCoreTesting/MetadataProviders/MockFrameworkMetadataProvider.swift
@@ -35,25 +35,3 @@ public final class MockFrameworkMetadataProvider: MockPrecompiledMetadataProvide
         }
     }
 }
-
-public extension FrameworkMetadata {
-    static func test(
-        path: AbsolutePath = "/Frameworks/TestFramework.xframework",
-        binaryPath: AbsolutePath = "/Frameworks/TestFramework.xframework/TestFramework",
-        dsymPath: AbsolutePath? = nil,
-        bcsymbolmapPaths: [AbsolutePath] = [],
-        linking: BinaryLinking = .dynamic,
-        architectures: [BinaryArchitecture] = [.arm64],
-        isCarthage: Bool = false
-    ) -> FrameworkMetadata {
-        FrameworkMetadata(
-            path: path,
-            binaryPath: binaryPath,
-            dsymPath: dsymPath,
-            bcsymbolmapPaths: bcsymbolmapPaths,
-            linking: linking,
-            architectures: architectures,
-            isCarthage: isCarthage
-        )
-    }
-}
